#69 rename: getProfileList 파일 이동, 폴더명 소문자로 변경
