+++ pages/wiki/[code]/[code].module.scss
@@ -1,4 +1,4 @@
-@import '../../styles/mixins';
+@import '../../../styles/mixins';
 
 .pcProfile {
   display: flex;
